+++ test/Feature/CleanRegex/Match/remaining/_stream/MatchPatternTest.php
@@ -12,10 +12,10 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_fluent_asInt_all()
+    public function shouldGet_stream_asInt_all()
     {
         // when
-        $all = pattern('\d+')->match('18 19')->remaining(Functions::equals('19'))->fluent()->asInt()->all();
+        $all = pattern('\d+')->match('18 19')->remaining(Functions::equals('19'))->stream()->asInt()->all();
 
         // then
         $this->assertSame([19], $all);
@@ -24,10 +24,10 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_fluent_keys_all()
+    public function shouldGet_stream_keys_all()
     {
         // when
-        $keys = pattern('\d+')->match('18 19')->remaining(Functions::equals('19'))->fluent()->keys()->all();
+        $keys = pattern('\d+')->match('18 19')->remaining(Functions::equals('19'))->stream()->keys()->all();
 
         // then
         $this->assertSame([0], $keys);
@@ -36,10 +36,10 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_fluent_keys_first()
+    public function shouldGet_stream_keys_first()
     {
         // when
-        $keys = pattern('\d+')->match('18 19 20')->remaining(Functions::equals('20'))->fluent()->keys()->first();
+        $keys = pattern('\d+')->match('18 19 20')->remaining(Functions::equals('20'))->stream()->keys()->first();
 
         // then
         $this->assertSame(0, $keys);
@@ -48,7 +48,7 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_offsets_fluent_keys_first()
+    public function shouldGet_offsets_keys_first()
     {
         // given
         $firstKey = pattern('\w+')->match('One Two Three')
@@ -64,7 +64,7 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_offsets_fluent_keys_all()
+    public function shouldGet_offsets_keys_all()
     {
         // given
         $keys = pattern('\w+')->match('One Two Three')